Update to operatorkit v1.0.0
