+++ go.mod
@@ -27,14 +27,14 @@ require (
 	github.com/giantswarm/ipam v0.2.0
 	github.com/giantswarm/k8sclient v0.2.0
 	github.com/giantswarm/k8sclient/v2 v2.0.0
-	github.com/giantswarm/k8sclient/v3 v3.0.0
+	github.com/giantswarm/k8sclient/v3 v3.1.0
 	github.com/giantswarm/k8scloudconfig/v6 v6.1.1
 	github.com/giantswarm/kubelock v0.2.0
 	github.com/giantswarm/microendpoint v0.2.0
 	github.com/giantswarm/microerror v0.2.0
 	github.com/giantswarm/microkit v0.2.0
 	github.com/giantswarm/micrologger v0.3.1
-	github.com/giantswarm/operatorkit v0.2.0
+	github.com/giantswarm/operatorkit v1.0.0
 	github.com/giantswarm/randomkeys v0.2.0
 	github.com/giantswarm/statusresource v0.3.0
 	github.com/giantswarm/tenantcluster v0.2.0
